# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ActorEventRead": {
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "first_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "reported_name": {
            "title": "Reported Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "summary",
          "status",
          "confidence",
          "first_seen",
          "last_seen",
          "reported_name"
        ],
        "title": "ActorEventRead",
        "type": "object"
      },
      "ActorTimelineBucket": {
        "properties": {
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "event_count"
        ],
        "title": "ActorTimelineBucket",
        "type": "object"
      },
      "ActorTrackingChangeSet": {
        "properties": {
          "category": {
            "enum": [
              "malware",
              "infrastructure",
              "techniques",
              "targets"
            ],
            "title": "Category",
            "type": "string"
          },
          "current_values": {
            "items": {
              "type": "string"
            },
            "title": "Current Values",
            "type": "array"
          },
          "disappeared_values": {
            "items": {
              "type": "string"
            },
            "title": "Disappeared Values",
            "type": "array"
          },
          "new_values": {
            "items": {
              "type": "string"
            },
            "title": "New Values",
            "type": "array"
          },
          "previous_values": {
            "items": {
              "type": "string"
            },
            "title": "Previous Values",
            "type": "array"
          }
        },
        "required": [
          "category",
          "current_values",
          "previous_values",
          "new_values",
          "disappeared_values"
        ],
        "title": "ActorTrackingChangeSet",
        "type": "object"
      },
...
      "ActorTrackingRead": {
        "properties": {
          "actor_id": {
            "format": "uuid",
            "title": "Actor Id",
            "type": "string"
          },
          "canonical_name": {
            "title": "Canonical Name",
            "type": "string"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/ActorTrackingChangeSet"
            },
            "title": "Changes",
            "type": "array"
          },
          "comparison": {
            "$ref": "#/components/schemas/ActorTrackingComparison"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ActorEventRead"
            },
            "title": "Events",
            "type": "array"
          },
          "period": {
            "$ref": "#/components/schemas/ActorTrackingPeriod"
          },
          "trend": {
            "items": {
              "$ref": "#/components/schemas/ActorTimelineBucket"
            },
            "title": "Trend",
            "type": "array"
          }
        },
        "required": [
          "actor_id",
          "canonical_name",
          "period",
          "comparison",
          "trend",
          "changes",
          "events"
        ],
        "title": "ActorTrackingRead",
        "type": "object"
      },
      "ActorTrackingSummaryRead": {
        "properties": {
          "actor_id": {
            "format": "uuid",
            "title": "Actor Id",
            "type": "string"
          },
          "caveats": {
            "items": {
              "type": "string"
            },
            "title": "Caveats",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "highlights": {
            "items": {
              "type": "string"
            },
            "title": "Highlights",
            "type": "array"
          },
          "method_version": {
            "default": "tracking-rules-v1",
            "title": "Method Version",
            "type": "string"
          },
          "status": {
            "const": "draft",
            "default": "draft",
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "supporting_event_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Supporting Event Ids",
            "type": "array"
          },
          "supporting_evidence_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Supporting Evidence Ids",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "title",
          "summary",
          "highlights",
          "caveats",
          "supporting_event_ids",
          "supporting_evidence_ids",
          "generated_at"
        ],
        "title": "ActorTrackingSummaryRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ThreatActorDetail": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "canonical_name": {
            "title": "Canonical Name",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ActorEventRead"
            },
            "title": "Events",
            "type": "array"
          },
          "first_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "latest_event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Event Id"
          },
          "latest_event_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Event Title"
          },
          "origin_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Country"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/ActorTimelineBucket"
            },
            "title": "Timeline",
            "type": "array"
          }
        },
        "required": [
          "id",
          "canonical_name",
          "aliases",
          "origin_country",
          "event_count",
          "first_seen",
          "last_seen",
          "latest_event_id",
          "latest_event_title",
          "description",
          "events",
          "timeline"
        ],
        "title": "ThreatActorDetail",
        "type": "object"
      },
      "ThreatActorSummary": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "canonical_name": {
            "title": "Canonical Name",
            "type": "string"
          },
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          },
          "first_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "latest_event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Event Id"
          },
          "latest_event_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Event Title"
          },
          "origin_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Country"
          }
        },
        "required": [
          "id",
          "canonical_name",
          "aliases",
          "origin_country",
          "event_count",
          "first_seen",
          "last_seen",
          "latest_event_id",
          "latest_event_title"
        ],
        "title": "ThreatActorSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "APT Hunter API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/actors": {
      "get": {
        "operationId": "list_threat_actors_api_v1_actors_get",
        "parameters": [
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ThreatActorSummary"
                  },
                  "title": "Response List Threat Actors Api V1 Actors Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Threat Actors",
        "tags": [
          "actors"
        ]
      }
    },
    "/api/v1/actors/{actor_id}": {
      "get": {
        "operationId": "get_threat_actor_api_v1_actors__actor_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "actor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Actor Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "granularity",
            "required": false,
            "schema": {
              "default": "month",
              "enum": [
                "month",
                "year"
              ],
              "title": "Granularity",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreatActorDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Threat Actor",
        "tags": [
          "actors"
        ]
      }
    },
    "/api/v1/actors/{actor_id}/tracking": {
      "get": {
        "operationId": "get_actor_tracking_api_v1_actors__actor_id__tracking_get",
        "parameters": [
          {
            "in": "path",
            "name": "actor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Actor Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "bucket",
            "required": false,
            "schema": {
              "default": "auto",
              "enum": [
                "auto",
                "day",
                "week",
                "month"
              ],
              "title": "Bucket",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorTrackingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Actor Tracking",
        "tags": [
          "actors"
        ]
      }
    },
    "/api/v1/actors/{actor_id}/tracking/export": {
      "get": {
        "operationId": "export_actor_tracking_api_v1_actors__actor_id__tracking_export_get",
        "parameters": [
          {
            "in": "path",
            "name": "actor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Actor Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "bucket",
            "required": false,
            "schema": {
              "default": "auto",
              "enum": [
                "auto",
                "day",
                "week",
                "month"
              ],
              "title": "Bucket",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Actor Tracking",
        "tags": [
          "actors"
        ]
      }
    },
    "/api/v1/actors/{actor_id}/tracking/summary": {
      "post": {
        "operationId": "generate_actor_tracking_summary_api_v1_actors__actor_id__tracking_summary_post",
        "parameters": [
          {
            "in": "path",
            "name": "actor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Actor Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "bucket",
            "required": false,
            "schema": {
              "default": "auto",
              "enum": [
                "auto",
                "day",
                "week",
                "month"
              ],
              "title": "Bucket",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorTrackingSummaryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Actor Tracking Summary",
        "tags": [
          "actors"
        ]
      }
    },
...
  }
}